Universal Viewer should always render via Javascript.
We have several complicated use cases around permissions and when a user
may possibly get access to material which can only be resolved by
querying Figgy to see if a user has access.

For OARSC we want the manifests to be indexed (so that a "online access"
badge appears), but we also want users to see the nice reading room
access click-through - not a giant white box where the UV should be.
Rather than adjusting the DAO and reindexing a collection every time a
component's permissions change we can just always check via our viewer
JS.

A side benefit is it gets rid of a bunch of renderer code.

Refs #948